requirements.txt
google-api-python-client==2.108.0
selenium==4.15.2
youtube-transcript-api==0.6.1
pytubefix==6.0.2
pandas==2.1.3
python-dotenv==1.0.0  